+++ src/extension.ts
@@ -53,13 +53,13 @@ export function activate(context: vscode.ExtensionContext) {
                 let comment = '';
                 if( pos.line > 0 ) { // dont look for comments if were at the top of the file
                     let offset = 1;
-                    let inBlock = false
+                    let inBlock = false;
                     // while not at the top of the file and the next line up is a comment
                     while(pos.line - offset >= 0) {
-                        let lineNum = pos.line - offset
+                        let lineNum = pos.line - offset;
                         let line = document.lineAt(new vscode.Position(lineNum, 0)).text;
-                        line = line.trimStart()
-                        if(offset == 1) { // if were right above the function declaration, look for a block comment
+                        line = line.trimStart();
+                        if(offset === 1) { // if were right above the function declaration, look for a block comment
                             let isBlock = G_BLOCK_COMMENTS_END.test(line);
                             G_BLOCK_COMMENTS_END.lastIndex = 0;
                             if(isBlock) {
@@ -79,7 +79,7 @@ export function activate(context: vscode.ExtensionContext) {
                                 // TRIM END TOKEN
                                 trimmedLine = newLine.split("*/"); // use whats after the star token
                                 if( trimmedLine.length > 2 ) {
-                                    let firstEl = trimmedLine.pop()
+                                    let firstEl = trimmedLine.pop();
                                 }
                                 newLine = trimmedLine.join("").trimStart();
 
@@ -96,7 +96,7 @@ export function activate(context: vscode.ExtensionContext) {
                                 // TRIM END TOKEN (guaranteed to not have text after end token if tthe user wants comments to appear in hover box)
                                 let trimmedLine = line.split("*/"); // use whats after the end token
                                 if( trimmedLine.length > 2 ) {
-                                    let firstEl = trimmedLine.pop()
+                                    let firstEl = trimmedLine.pop();
                                 }
                                 let newLine = trimmedLine.join("").trimStart();
 
@@ -196,7 +196,7 @@ function newHoverText(key: string, text: string, docUrl: string, ...args: string
         let codeBlock = false;
         for (let i = 0; i < commentLines.length; i++) {
             let line = commentLines[i].replace(/^\/\//g, "");
-            line = line.trimStart()
+            line = line.trimStart();
             if (line.startsWith('```')) {
                 codeBlock = !codeBlock;
                 if(codeBlock) {